Feat/#69: 예약 내역에 대한 평점 및 후기 남기기

+++ app/src/main/java/org/sopar/data/remote/response/ParkingLot.kt
@@ -36,7 +36,7 @@ data class ParkingLot(
     @field:Json(name = "phoneNumber")
     val phoneNumber: String,
     @field:Json(name = "ratingAvg")
-    val ratingAvg: Int?,
+    val ratingAvg: Float?,
     @field:Json(name = "ratingNum")
     val ratingNum: Int?,
     @field:Json(name = "remainingSpace")
